Vector: change id format client (#3817)
[EP-465] client side.

Deprecating `name` turned out to be a little hairy.  There are two classes of deprecation going on for `name`
* `name` really meant `product_id` (like in `FeatureCollection.create`), I just renamed the arg to `product_id` and collapse kwarg calls like `name="some-name"` to `product_id="some-name"` raising exceptions on clashes.
* `name` actually meant `name` (like in `FeatureCollection.replace`) and the value is no longer required, because `name` will go away eventually.  In these cases removing the parameter from the function completely is desired, but can't happen without breaking code; so I transition args to kwargs with defaults to make `name` optional and indicating which of the other following kwargs are required.

I added the `deprecate` decorator in `deprecation.py` to help with renaming and eventually removing parameters in client code.  The functionality of `deprecate` is limited to this use case, wondering if this is the right approach for this, and if this is a useful thing moving forward.

[EP-465] #done

GitOrigin-RevId: f2892630446046e70034caad3d931270946d08a8